Added NewEmptyResouce and unexported builtin detectors
